chore: demog mode一覧でログ一覧画面に遷移可能にした
